+++ trunk/ libnoise-dotnet/libnoise/src/IModule4D.cs
@@ -35,7 +35,7 @@ namespace Graphics.Tools.Noise {
 		/// <param name="z">The input coordinate on the z-axis.</param>
 		/// <param name="z">The input coordinate on the t-axis.</param>
 		/// <returns>The resulting output value.</returns>
-		double GetValue(double x, double y, double z, double t);
+		float GetValue(float x, float y, float z, float t);
 
 
 		#endregion
